Added file for gui layout, assembly machine inventory rendering
More work is required for the progress bar in the assembly machine Gui to line up with the adjacent slot.

Bugfixes:
- Fixed selected recipe category not being highlighted
- Fixed gui positioning issues after using DrawSlots()

Sometimes when the ImGui draw cursor is positioned with SetCursorPos(x, y), ImGui decides to change where the actual widget is drawn with little explanation.

For example, if the last slot is not drawn when using DrawSlots(), the y position of the next widget is at the position of the previous widget or far below where it should have been.

To mitigate some of the issues, DrawItemSlot() no longer affects the draw cursor position. Only DrawSlots(), which upon exiting resets the draw cursor position for the next widget (usually).