Fixed multi tile entities not being deactivated upon being picked up
To prevent attempting to render a gui for a non existent entity, if the player has a gui open for an entity and they pick the entity up, their activatedLayer will be unset.

This had not worked for multi tiles if the player activates and removes on different tiles. It was fixed by always activating and checking the top left tile prior to removal.