Fixed update dispatcher unregister invalid read
The update dispatcher groups callbacks by game tick, when a
callback is removed and its group is empty, the group is deleted.

However, to exit a for loop iterating over the group to delete the
callback, it checked the group's size with a now invalid reference.
(since it was deleted)